chore: deduplication helper time parsing function
