+++ src/utils/mod.rs
@@ -1,9 +1,11 @@
+mod duration;
 mod fs_atomic;
 mod git;
 mod ids;
 mod json_object;
 mod plurals;
 
+pub use duration::parse_duration_to_seconds;
 pub use fs_atomic::write_text_atomic;
 pub use git::GitRepo;
 pub(crate) use ids::new_uuid_id;
